RUB-263 rust: split OpenSSL verify module
